Create Kmhmu' Alphabet
Khmu is a member of the Khmuic branch of the Austroasiatic language family. It is spoken mainly in northern Laos, and also in nearby parts of Vietnam, Thailand and China. In 2015 there were 708,000 speakers of Khmu in Laos. There were 72,900 speakers in Vietnam in 2009, 7,000 speakers in China in 2010, 6,250 speakers in Thailand in 2007, and in 2010 there were about 8,000 speakers of Khmu in the USA.
There are two groups of Khmu dialects: Western and Eastern. Dialects spoken in neighboring areas are more or less mutually intelligible, however speakers of more distant dialects may have difficulties communicating with each other. The written language is based on Eastern dialects.